LSW-52 Create exception subpackage and move files into it
