+++ modules/liferay-sample-module/liferay-sample-module-ws-client/src/main/java/com/liferay/sample/module/ws/exception/ApiResponseException.java
@@ -1,4 +1,4 @@
-package com.liferay.sample.module.ws.api;
+package com.liferay.sample.module.ws.exception;
 
 /**
  * @author Rafael Oliveira
